requirements.txt
python-telegram-bot==21.0
PyGithub==2.3.0
python-dotenv==1.0.0
